# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: MockAgents API
  version: 0.4.0
  description: >
    MockAgents is a platform for mocking AI agent integrations. It exposes a
    Management API (agents, interaction logs, costs, audit, pipelines, config
    validation, and a multi-tenant control plane of tenants + RBAC API keys),
    and Protocol Endpoints that mimic real LLM provider APIs (OpenAI Chat
    Completions and Anthropic Messages).


    The Model Context Protocol mock (`/mcp`, `/mcp/events`, `/mcp/response`) is
    JSON-RPC 2.0 rather than REST and is documented in the project README,
    not in this OpenAPI document.


    **Authentication.** When the server runs with `MOCKAGENTS_MULTI_TENANT=1`,
    every `/api/v1/*` management route requires an API key
    (`Authorization: Bearer <key>` or `X-Api-Key: <key>`) whose role meets the
    route's floor (`viewer < editor < admin < platform`). In single-tenant mode
    the management API is open. The protocol endpoints (`/v1/*`) are always
    unauthenticated — clients send their own provider key, which MockAgents
    ignores.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/pipelines:
    get:
      operationId: listPipelines
      summary: List pipelines
      description: "Returns all loaded `kind: Pipeline` documents."
      tags: [Pipelines]
      responses:
        "200":
          description: Pipelines.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Pipeline" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/pipelines/{name}:
    get:
      operationId: getPipeline
      summary: Get a pipeline (with DAG)
      tags: [Pipelines]
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Pipeline detail including nodes and edges. The `ETag` response header carries the version to echo back as `If-Match` on update.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Pipeline" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      operationId: updatePipeline
      summary: Update a pipeline (write-back to disk)
      description: >
        Validate an edited Pipeline definition and, on success, persist it
        atomically to its source file and re-register it so the change is live
        immediately. Requires `editor` in multi-tenant mode. Optimistic
        concurrency: send the `ETag` from GET as `If-Match`; the write is
        rejected if the pipeline changed on disk since it was loaded.
      tags: [Pipelines]
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
        - name: If-Match
          in: header
          required: true
          description: The `ETag` returned by GET. Required for optimistic concurrency.
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Pipeline" }
      responses:
        "200":
          description: Saved. The new version is returned in the `ETag` response header.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Pipeline" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "412":
          description: The `If-Match` version no longer matches; the pipeline changed on disk. Reload and re-apply.
        "422":
          description: The edited definition failed validation; the file was not written.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidateResult" }
        "428":
          description: The `If-Match` header is missing.
...
components:
...
  schemas:
...
    Error:
      type: object
      description: >
        The management API's error envelope. Every 4xx and 5xx from a
        /api/v1/... route has this shape.

        NOT what the provider endpoints return when a quota rejects a request:
        those mimic the upstream `{"error": {"type", "message"}}` object so an
        SDK surfaces them correctly. See ProviderQuotaError.
      required: [error]
      properties:
        error:
          type: string
          example: agent "weather-agent" not found
        available_agents:
          type: array
          items: { type: string }
          description: >
            Present on a 404 for an unknown agent — what the caller could have
            asked for instead. Tenant-scoped: it never names another tenant's
            agents.
        details:
          type: string
          description: >
            Present when a write or reload failed schema validation; carries the
            validator's output.
...
    Pipeline:
      type: object
      description: >
        A pipeline document, in the same envelope the YAML uses. As with
        AgentDefinition, only the envelope is described here: the contents of
        `spec` are defined by the loader and duplicating them would create a
        second definition to keep in sync.
      required: [apiVersion, kind, metadata, spec]
      properties:
        apiVersion:
          type: string
          example: mockagents/v1
        kind:
          type: string
          enum: [Pipeline]
        metadata:
          type: object
          additionalProperties: true
        spec:
          type: object
          description: >
            `topology` (sequential | parallel | graph), `agents` (node id and
            agent ref) and, for a graph, `edges` with optional `when_contains`.
          additionalProperties: true
    ValidateResult:
      type: object
      required: [ok, kind, errors]
      properties:
        ok: { type: boolean }
        kind: { type: string, example: Agent }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
              line: { type: integer }
              suggestion: { type: string }

  # ---------------------------------------------------------------------------
  # Reusable Responses
  # ---------------------------------------------------------------------------
  responses:
    BadRequest:
      description: The request was malformed or contained invalid parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: bad_request
              message: "Invalid value for parameter 'limit': must be a positive integer."
              details: {}
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: not_found
              message: The requested resource was not found.
              details: {}
...
    Unauthorized:
      description: Missing or invalid API key (multi-tenant mode).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              message: missing Authorization bearer token or X-Api-Key header
              type: authentication_error
    Forbidden:
      description: Authenticated, but the key's role is below the route's floor.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        `Authorization: Bearer <api-key>`. Required on `/api/v1/*` routes when
        the server runs with `MOCKAGENTS_MULTI_TENANT=1`.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: "Alternative to `Authorization: Bearer` for the same API key."

# Applies to the management API in multi-tenant mode. Protocol endpoints
# (/v1/*) are always unauthenticated and override this with `security: []`.
security:
  - BearerAuth: []
  - ApiKeyAuth: []